#231: move compiler_options_t to a separate utility file
